docker: split reconciler.rs into sibling modules (review #9)
Behavior-preserving, mechanical extraction to address the reconciler's size
(review code-quality finding). reconciler.rs 4032 → 1769 lines (−2263); the whole
`impl DockerReconciler` async block stays put — only &self-free items + their
tests moved:
- diff.rs    — ActualContainer, ReconcileAction, diff_desired_vs_actual,
               identity_key/spec_key/action_key + diff/scale tests
- rolling.rs — filter_rolling_actions, routable_for, rolling_rotation_decision,
               replica_ready/ReadyVerdict, service_names_for_spec + tests
- health.rs  — effective_health_path, probe_error_detail + tests
- env.rs      — merge_container_env, upsert_env, hash_env + tests
- pod_status.rs — build_controller_metadata/build_pod_status + tests
- test_helpers.rs — shared #[cfg(test)] fixtures used by diff/rolling tests

Cross-module helpers tightened to pub(crate); a stray merge_container_env
doc-comment that had drifted onto probe_error_detail was put back. No logic,
signature, or behavior change. 120 docker-controller + 35 settings tests pass,
clippy clean.
