+++ src/server/deployment/controller/docker/reconciler.rs
@@ -15,12 +15,18 @@ use anyhow::Result;
 use bollard::Docker;
 use chrono::Utc;
 use futures::StreamExt;
-use sha2::{Digest, Sha256};
 use sqlx::PgPool;
 use tracing::{debug, error, info, warn};
 
 use super::container_builder::{self, BuilderConfig, DesiredContainer, DesiredRoute};
+use super::diff::{diff_desired_vs_actual, identity_key, ActualContainer, ReconcileAction};
+use super::env::{hash_env, merge_container_env, upsert_env};
+use super::health::{effective_health_path, probe_error_detail};
 use super::labels::{self, SUFFIX_ENV_HASH, SUFFIX_IMAGE, SUFFIX_MANAGED_BY};
+use super::pod_status::build_controller_metadata;
+use super::rolling::{
+    filter_rolling_actions, replica_ready, routable_for, service_names_for_spec, ReadyVerdict,
+};
 use crate::db::models::{Deployment, DeploymentStatus, Project, TerminationReason};
 use crate::server::deployment::models::rise_system_env_vars;
 use crate::server::deployment::resource_builder::ResourceBuilder;
@@ -1703,530 +1709,6 @@ impl DockerReconciler {
     }
 }
 
-// ── Pure pod_status builder ───────────────────────────────────────────────
-
-/// Assemble the `controller_metadata` JSON for the deployment: a
-/// K8s-pod-status-shaped `pod_status` block (so the Pods tab renders unchanged —
-/// see `frontend/src/features/deployments.tsx`) plus the legacy sibling `health`
-/// block. `pods` is one (name, inspection) entry PER REPLICA container across all
-/// specs (so `desired_replicas` = sum of N over specs, `current_replicas` =
-/// running count, `ready_replicas` = healthy count). Pure so it can be
-/// unit-tested without a daemon. The container readiness verdict (`is_ready`) is
-/// the same one driving the status transitions.
-fn build_controller_metadata(
-    pods: &[(String, Option<InspectedContainer>)],
-    status: &DeploymentStatus,
-    is_ready: bool,
-) -> serde_json::Value {
-    let pod_status = build_pod_status(pods, is_ready);
-    let desired = pods.len();
-    let ready_replicas = pod_status
-        .get("ready_replicas")
-        .and_then(|v| v.as_u64())
-        .unwrap_or(0) as usize;
-
-    serde_json::json!({
-        "pod_status": pod_status,
-        "health": {
-            "last_check": Utc::now().to_rfc3339(),
-            // Healthy only when EVERY replica of EVERY spec is ready AND the
-            // deployment isn't currently flagged Unhealthy.
-            "healthy": ready_replicas == desired
-                && *status != DeploymentStatus::Unhealthy,
-        },
-    })
-}
-
-/// Build the K8s-pod-status-shaped `pod_status` JSON from each container's
-/// (name, inspection). Pure so it can be unit-tested without a daemon: one
-/// container == one pod (as the Docker controller has always modeled it).
-///
-/// Mirrors the shape produced by the K8s path (`webhook.rs`) so the existing
-/// frontend (`frontend/src/features/deployments.tsx`), API and DB need no
-/// changes:
-///   - top-level: `desired_replicas`, `current_replicas` (running), `ready_replicas`,
-///     `pods`, `last_checked`.
-///   - per pod: `name`, `phase` (running→Running, created/restarting→Pending,
-///     exited/dead→Failed, else Unknown), `conditions: []`, `containers: [...]`.
-///   - per container: `name`, `ready`, `restart_count`, `state` with
-///     `state_type` (running|waiting|terminated), `started_at`, `finished_at`,
-///     `exit_code`, `reason`.
-///
-/// `is_ready` is the reconciler's overall readiness verdict (every container
-/// must be ready). A container's per-pod `ready` reflects that verdict only when
-/// the container is actually running, so a crashed container never shows ready.
-fn build_pod_status(
-    named: &[(String, Option<InspectedContainer>)],
-    is_ready: bool,
-) -> serde_json::Value {
-    let mut pods = Vec::with_capacity(named.len());
-    let mut current_replicas = 0usize; // running
-    let mut ready_replicas = 0usize;
-
-    for (name, inspected) in named {
-        let status = inspected
-            .as_ref()
-            .and_then(|i| i.status.as_deref())
-            .unwrap_or("");
-        let running = inspected.as_ref().map(|i| i.running).unwrap_or(false);
-        if running {
-            current_replicas += 1;
-        }
-        // A container is "ready" for the pod view when the deployment is ready
-        // overall AND this container is running.
-        let container_ready = is_ready && running;
-        if container_ready {
-            ready_replicas += 1;
-        }
-
-        let phase = match status {
-            "running" => "Running",
-            "created" | "restarting" => "Pending",
-            "exited" | "dead" => "Failed",
-            _ => "Unknown",
-        };
-
-        let restart_count = inspected
-            .as_ref()
-            .and_then(|i| i.restart_count)
-            .unwrap_or(0);
-        let started_at = inspected.as_ref().and_then(|i| i.started_at.clone());
-        let finished_at = inspected.as_ref().and_then(|i| i.finished_at.clone());
-        let exit_code = inspected.as_ref().and_then(|i| i.exit_code);
-        let error = inspected.as_ref().and_then(|i| i.error.clone());
-        let health = inspected.as_ref().and_then(|i| i.health.clone());
-
-        // Map the Docker state to the K8s container-state shape the frontend's
-        // `ContainerState` interface expects.
-        let state = match status {
-            "running" => serde_json::json!({
-                "state_type": "running",
-                "started_at": started_at,
-                "finished_at": serde_json::Value::Null,
-                "exit_code": serde_json::Value::Null,
-                // Surface a Docker HEALTHCHECK verdict (when an image ships one)
-                // as the running-state reason — e.g. "starting"/"unhealthy".
-                // `none`/`healthy` (the common case; Rise injects no HEALTHCHECK)
-                // leave the reason null.
-                "reason": health
-                    .filter(|h| h != "none" && h != "healthy"),
-            }),
-            "exited" | "dead" => serde_json::json!({
-                "state_type": "terminated",
-                "started_at": started_at,
-                "finished_at": finished_at,
-                "exit_code": exit_code,
-                // Prefer the daemon's error string; otherwise synthesize from
-                // the exit code so a non-zero exit still has a visible reason.
-                "reason": error.clone().or_else(|| {
-                    exit_code.map(|c| if c == 0 { "Completed".to_string() } else { format!("Error (exit {c})") })
-                }),
-            }),
-            // created / restarting / unknown / missing → waiting.
-            _ => serde_json::json!({
-                "state_type": "waiting",
-                "started_at": serde_json::Value::Null,
-                "finished_at": serde_json::Value::Null,
-                "exit_code": serde_json::Value::Null,
-                "reason": if status.is_empty() { "ContainerCreating".to_string() } else { status.to_string() },
-            }),
-        };
-
-        pods.push(serde_json::json!({
-            "name": name,
-            "phase": phase,
-            "conditions": [],
-            "containers": [{
-                "name": name,
-                "ready": container_ready,
-                "restart_count": restart_count,
-                "state": state,
-            }],
-        }));
-    }
-
-    serde_json::json!({
-        "desired_replicas": named.len(),
-        "current_replicas": current_replicas,
-        "ready_replicas": ready_replicas,
-        "pods": pods,
-        "last_checked": Utc::now().to_rfc3339(),
-    })
-}
-
-// ── Pure diff ───────────────────────────────────────────────────────────
-
-/// Observed Rise-managed container, reduced to the fields the diff needs.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ActualContainer {
-    pub id: String,
-    pub name: String,
-    /// The `project` bookkeeping label. Part of the stable identity tuple used
-    /// to match this container to a desired one across generations (the live
-    /// `--name` carries a per-recreate `_g{n}` suffix and can't be matched
-    /// directly). `None` for legacy containers missing the label.
-    pub project: Option<String>,
-    /// The `deployment-group` bookkeeping label (identity tuple member).
-    pub deployment_group: Option<String>,
-    /// The `container` bookkeeping label (identity tuple member).
-    pub container: Option<String>,
-    /// The `deployment-id` bookkeeping label stamped on the live container.
-    /// Used by the diff to protect a container from GC when its owning
-    /// deployment's desired computation failed this tick, and as a member of the
-    /// stable identity tuple. `None` for legacy containers created before this
-    /// label was read (treated as unprotected).
-    pub deployment_id_label: Option<String>,
-    /// Monotonic generation parsed from the `generation` label. The NEXT
-    /// generation on a recreate is `generation + 1`. Defaults to `0` when the
-    /// label is missing/legacy/unparseable, so the first recreate of a legacy
-    /// container yields generation `1`.
-    pub generation: u32,
-    /// Zero-based replica index parsed from the `replica` label. Member of the
-    /// stable identity tuple, so each replica is matched/recreated independently.
-    /// Defaults to `0` when the label is missing/legacy/unparseable, so a
-    /// pre-replica container maps onto replica 0 of its spec.
-    pub replica: u32,
-    pub image_label: Option<String>,
-    pub env_hash_label: Option<String>,
-    /// The `route-hash` recreate-signature label stamped on the live container
-    /// (sha256 of its rendered Traefik label set plus its published-port intent).
-    /// `None` for legacy containers created before this label existed. Compared
-    /// against the desired container's `route_hash` so a routing transition (a
-    /// deployment becoming or ceasing to be active) OR a published-port change
-    /// forces a recreate — Docker can't mutate a running container's labels or
-    /// port bindings in place.
-    pub route_hash_label: Option<String>,
-    /// Daemon-reported lifecycle state (e.g. "running", "exited", "created",
-    /// "dead", "restarting", "paused"). `None` when the daemon didn't report
-    /// it. A container that matches on image + env but is in a terminal-ish
-    /// state (`exited`/`created`/`dead`/`None`) is recreated; transient managed
-    /// states (`running`/`restarting`/`paused`) are left alone — see
-    /// `ActualContainer::needs_recreate`.
-    pub state: Option<String>,
-}
-
-impl ActualContainer {
-    /// Stable identity of this container from its bookkeeping LABELS — the tuple
-    /// (project, deployment_group, deployment_id, container). Used to match an
-    /// actual container to a desired one WITHOUT relying on the live `--name`,
-    /// which now carries a per-recreate `_g{n}` generation suffix. Returns
-    /// `None` when any identity label is missing (a legacy/foreign container
-    /// that can't be matched — it's only ever a removable orphan, if not
-    /// protected).
-    fn identity(&self) -> Option<String> {
-        Some(identity_key(
-            self.project.as_deref()?,
-            self.deployment_group.as_deref()?,
-            self.deployment_id_label.as_deref()?,
-            self.container.as_deref()?,
-            self.replica,
-        ))
-    }
-
-    /// Spec-level key (identity minus replica) for grouping replicas of a spec.
-    /// `None` when any spec-identity label is missing (a legacy/foreign
-    /// container that can't be grouped).
-    fn spec_identity(&self) -> Option<String> {
-        Some(spec_key(
-            self.project.as_deref()?,
-            self.deployment_group.as_deref()?,
-            self.deployment_id_label.as_deref()?,
-            self.container.as_deref()?,
-        ))
-    }
-
-    /// Whether the daemon reports this container in a terminal-ish state that
-    /// won't self-recover, so the reconciler must recreate it.
-    ///
-    /// Only `exited`/`created`/`dead` (and an unreported state, `None`) qualify:
-    /// these never start serving again on their own. Transient daemon-managed
-    /// states — `running`, `restarting` (Docker's restart policy applying
-    /// backoff after a crash) and `paused` (operator-initiated) — must NOT be
-    /// force-recreated: doing so every reconcile tick would reset Docker's
-    /// restart history and fight its exponential backoff. The health probe
-    /// drives the Unhealthy status independently, so a crash-looping container
-    /// is reflected in status without the reconciler churning it.
-    fn needs_recreate(&self) -> bool {
-        matches!(
-            self.state.as_deref(),
-            Some("exited") | Some("created") | Some("dead") | None
-        )
-    }
-}
-
-/// What to do with one container slot.
-///
-/// Create/Recreate carry the stable `identity` (so `apply_actions` can find the
-/// desired container by its identity tuple), the resolved generation-ful `name`
-/// (for logging/clarity), and the `generation` to stamp on the new container.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ReconcileAction {
-    /// No matching actual container — create + start it at `generation`.
-    Create {
-        identity: String,
-        name: String,
-        generation: u32,
-    },
-    /// Actual exists but drifted (image / env-hash / route-hash / not running)
-    /// — replace it, bumping to `generation` (live generation + 1).
-    Recreate {
-        identity: String,
-        name: String,
-        existing_id: String,
-        generation: u32,
-    },
-    /// Actual exists with no desired match — garbage-collect it.
-    Remove { id: String, name: String },
-}
-
-/// Classify each desired/actual container into create / recreate / remove.
-///
-/// Pure: no daemon access. Matching is by the STABLE IDENTITY tuple (project,
-/// deployment_group, deployment_id, container) read from bookkeeping labels —
-/// NOT by container name, which now carries a per-recreate `_g{n}` generation
-/// suffix. `container_prefix` is the configured prefix used only to render the
-/// resolved generation-ful `name` carried on each Create/Recreate action (for
-/// logging); `apply_actions` looks the desired container up by `identity`.
-///
-/// `protected_deployment_ids` is the set of deployment-ids whose desired
-/// computation FAILED this tick. An actual container whose `deployment-id`
-/// label is in this set is never classified as `Remove`, even though it has no
-/// desired match — its absence from `desired` is an artifact of the transient
-/// failure, not a genuine orphan. This keeps healthy, actively-serving
-/// containers running across a brief KMS/DB hiccup, matching the K8s model
-/// (children untouched on a failed sync). Containers for deployments that
-/// successfully computed and are genuinely orphaned still get GC'd.
-pub fn diff_desired_vs_actual(
-    desired: &[DesiredContainer],
-    actual: &[ActualContainer],
-    container_prefix: &str,
-    protected_deployment_ids: &HashSet<String>,
-) -> Vec<ReconcileAction> {
-    // Desired containers keyed by their STABLE IDENTITY tuple (project, group,
-    // deployment-id, container) — NOT by their live `--name`, which now carries
-    // a per-recreate `_g{n}` generation suffix that differs between an actual
-    // container and the next-generation name we'd compute for it.
-    let mut desired_by_identity: HashMap<String, &DesiredContainer> = HashMap::new();
-    for d in desired {
-        let key = identity_key(
-            &d.project,
-            &d.deployment_group,
-            &d.deployment_id,
-            &d.container,
-            d.replica,
-        );
-        desired_by_identity.insert(key, d);
-    }
-
-    // Actual containers keyed by the identity tuple read from their bookkeeping
-    // labels. Containers missing identity labels (`identity() == None`) are not
-    // indexed here — they can never match a desired slot and are only ever
-    // removable orphans (handled in the GC pass below).
-    let actual_by_identity: HashMap<String, &ActualContainer> = actual
-        .iter()
-        .filter_map(|a| a.identity().map(|id| (id, a)))
-        .collect();
-
-    let mut actions = Vec::new();
-    let mut matched_actual_ids: HashSet<String> = HashSet::new();
-
-    for (identity, d) in &desired_by_identity {
-        match actual_by_identity.get(identity) {
-            None => {
-                // Brand-new slot: create at generation 1.
-                actions.push(ReconcileAction::Create {
-                    identity: identity.clone(),
-                    name: container_builder::container_name(
-                        container_prefix,
-                        &d.project,
-                        &d.deployment_group,
-                        &d.deployment_id,
-                        &d.container,
-                        d.replica,
-                        1,
-                    ),
-                    generation: 1,
-                });
-            }
-            Some(a) => {
-                matched_actual_ids.insert(a.id.clone());
-                let image_drift = a.image_label.as_deref() != Some(d.image.as_str());
-                let env_drift = a.env_hash_label.as_deref() != Some(d.env_hash.as_str());
-                // Recreate-signature drift: the live container's stamped
-                // `route-hash` differs from the desired one. This makes
-                // create-time-only properties a first-class part of the diff —
-                // when a deployment becomes active (gains Traefik labels) or
-                // stops being active (loses them), OR when its app port starts/
-                // stops being published to a loopback host port, the hash changes
-                // and forces a recreate to apply it. Docker can't mutate a running
-                // container's labels or port bindings in place, so recreation is
-                // the only way. Idempotent: once the stamped hash equals the
-                // desired one nothing recreates. A legacy container missing the
-                // label (`None`) is recreated once to gain it, then converges.
-                //
-                // NOTE: the generation is deliberately NOT part of any of these
-                // drift checks (it isn't fed into image/env/route hashes either),
-                // so a stable container produces ZERO actions regardless of its
-                // `_g{n}` suffix — the key no-infinite-recreate invariant.
-                let route_drift = a.route_hash_label.as_deref() != Some(d.route_hash.as_str());
-                // A container that matches on image + env but is in a
-                // terminal-ish state (created-but-never-started, exited, or
-                // crash-give-up `dead`) must be recreated so the deployment
-                // recovers. Transient managed states (`restarting`, `paused`)
-                // are left to Docker's restart policy so its backoff can settle.
-                let not_running = a.needs_recreate();
-                if image_drift || env_drift || route_drift || not_running {
-                    // Bump to the next generation so the replacement's name is
-                    // visibly newer than the one it replaces.
-                    let generation = a.generation + 1;
-                    actions.push(ReconcileAction::Recreate {
-                        identity: identity.clone(),
-                        name: container_builder::container_name(
-                            container_prefix,
-                            &d.project,
-                            &d.deployment_group,
-                            &d.deployment_id,
-                            &d.container,
-                            d.replica,
-                            generation,
-                        ),
-                        existing_id: a.id.clone(),
-                        generation,
-                    });
-                }
-            }
-        }
-    }
-
-    for a in actual {
-        if matched_actual_ids.contains(&a.id) {
-            continue;
-        }
-        // Protect containers belonging to a deployment whose desired
-        // computation failed this tick: treat them as matched rather than
-        // orphaned, so a transient error can't GC a healthy container.
-        if a.deployment_id_label
-            .as_deref()
-            .is_some_and(|id| protected_deployment_ids.contains(id))
-        {
-            continue;
-        }
-        actions.push(ReconcileAction::Remove {
-            id: a.id.clone(),
-            name: a.name.clone(),
-        });
-    }
-
-    // Deterministic ordering for testability.
-    actions.sort_by_key(action_key);
-    actions
-}
-
-/// Apply the ROLLING-RECREATE throttle to a diff's actions, run AFTER
-/// [`diff_desired_vs_actual`] and BEFORE `apply_actions`. Enforces, per spec
-/// (grouped by the spec key = identity minus replica), that a *running but
-/// drifted* replica is recreated at most ONE-AT-A-TIME and only while every
-/// OTHER replica of that spec is currently HEALTHY — so a rollout never drops
-/// more than one replica of capacity at once.
-///
-/// Pass-through rules (PER SPEC):
-/// - **Create** / **Remove** → ALWAYS pass (initial rollout / scale-up add
-///   capacity; scale-down / GC remove it). Never throttled.
-/// - **Recreate** of a recovery replica — the matched live container's `state`
-///   is NOT "running" (crashed / exited / created / missing) → ALWAYS pass; it's
-///   already down, so recreating only restores capacity.
-/// - **Recreate** of a rollout replica — the matched live container IS "running"
-///   but drifted → emit AT MOST ONE per spec this tick, the lowest replica index,
-///   and ONLY IF every OTHER replica of the spec is HEALTHY
-///   (`healthy_by_identity`). If ANY sibling is unhealthy/starting, defer ALL
-///   rollout recreates for the spec this tick.
-///
-/// INVARIANT: for a single spec with R running+drifted+healthy replicas this
-/// yields exactly ONE rollout Recreate; with any sibling unhealthy it yields
-/// ZERO; recovery Recreates and all Creates/Removes pass regardless.
-///
-/// `actual` provides each container's run state + identity (to map an action's
-/// identity → its spec + replica index + run state). `healthy_by_identity` maps
-/// an `identity_key(...)` to whether that live replica passed the HTTP health
-/// probe this tick (absent → treated as NOT healthy).
-fn filter_rolling_actions(
-    actions: Vec<ReconcileAction>,
-    actual: &[ActualContainer],
-    healthy_by_identity: &HashMap<String, bool>,
-) -> Vec<ReconcileAction> {
-    // identity → the live container (for run state + replica/spec grouping).
-    let actual_by_identity: HashMap<String, &ActualContainer> = actual
-        .iter()
-        .filter_map(|a| a.identity().map(|id| (id, a)))
-        .collect();
-
-    // Collect rollout-recreate candidates per spec; pass everything else through.
-    // A candidate is (replica_index, action) where the matched live container is
-    // running. We then admit at most one per spec, gated on sibling health.
-    let mut passed: Vec<ReconcileAction> = Vec::new();
-    // spec_key → Vec<(replica, identity, action)> of running+drifted recreates.
-    let mut rollout_candidates: HashMap<String, Vec<(u32, String, ReconcileAction)>> =
-        HashMap::new();
-
-    for action in actions {
-        match &action {
-            ReconcileAction::Create { .. } | ReconcileAction::Remove { .. } => {
-                // Unthrottled: capacity add (Create) / scale-down or GC (Remove).
-                passed.push(action);
-            }
-            ReconcileAction::Recreate { identity, .. } => {
-                match actual_by_identity.get(identity) {
-                    Some(a) if a.state.as_deref() == Some("running") => {
-                        // Rollout candidate: a running, drifted replica. Defer the
-                        // admission decision until we've seen all of them per spec.
-                        let key = a.spec_identity().unwrap_or_else(|| identity.clone());
-                        rollout_candidates.entry(key).or_default().push((
-                            a.replica,
-                            identity.clone(),
-                            action,
-                        ));
-                    }
-                    // Recovery (not running) OR no matched live container (already
-                    // gone) → pass unthrottled: recreating restores capacity.
-                    _ => passed.push(action),
-                }
-            }
-        }
-    }
-
-    // Admit at most one rollout recreate per spec, the lowest replica index, and
-    // only when every OTHER replica of the spec is healthy.
-    for (spec, mut candidates) in rollout_candidates {
-        candidates.sort_by_key(|(replica, _, _)| *replica);
-        // Identities of this spec's replicas slated for a rollout recreate — they
-        // don't count as "must be healthy" siblings (they're already drifted and
-        // about to be replaced), so a uniform drift across all replicas still
-        // rolls one-by-one rather than deadlocking.
-        let candidate_identities: HashSet<String> =
-            candidates.iter().map(|(_, id, _)| id.clone()).collect();
-
-        // Gate: every OTHER replica of this spec (a live replica NOT itself a
-        // rollout candidate) must be healthy. If any such sibling is
-        // unhealthy/starting, defer ALL rollout recreates for this spec.
-        let siblings_healthy = actual
-            .iter()
-            .filter(|a| a.spec_identity().as_deref() == Some(spec.as_str()))
-            .filter_map(|a| a.identity())
-            .filter(|id| !candidate_identities.contains(id))
-            .all(|id| healthy_by_identity.get(&id).copied().unwrap_or(false));
-
-        if siblings_healthy {
-            if let Some((_, _, action)) = candidates.into_iter().next() {
-                passed.push(action);
-            }
-        }
-        // else: defer ALL of this spec's rollout recreates this tick.
-    }
-
-    passed.sort_by_key(action_key);
-    passed
-}
-
 /// Pure comparison of this controller's configured class against the
 /// project's Organization's `deploymentControllerClass`. Mirrors the K8s
 /// webhook's `check_controller_class`:
@@ -2242,37 +1724,6 @@ pub fn controller_class_matches(configured: &str, org_class: Option<&str>) -> bo
     org_class == Some(configured)
 }
 
-fn action_key(a: &ReconcileAction) -> (u8, String) {
-    match a {
-        ReconcileAction::Create { name, .. } => (0, name.clone()),
-        ReconcileAction::Recreate { name, .. } => (1, name.clone()),
-        ReconcileAction::Remove { name, .. } => (2, name.clone()),
-    }
-}
-
-/// Stable identity key for a container slot, joining the identity tuple
-/// (project, deployment_group, deployment_id, container, replica) with a NUL
-/// separator (which can't appear in any of these names). Built from desired
-/// fields and from actual bookkeeping labels so the two match across name
-/// generations. The replica index is the matching unit: each replica of a spec
-/// is reconciled independently.
-fn identity_key(
-    project: &str,
-    group: &str,
-    deployment_id: &str,
-    container: &str,
-    replica: u32,
-) -> String {
-    format!("{project}\u{0}{group}\u{0}{deployment_id}\u{0}{container}\u{0}{replica}")
-}
-
-/// Spec-level key = the identity tuple MINUS the replica index
-/// (project, group, deployment_id, container). Groups all replicas of a single
-/// spec together — used by the rolling-recreate throttle to reason per-spec.
-fn spec_key(project: &str, group: &str, deployment_id: &str, container: &str) -> String {
-    format!("{project}\u{0}{group}\u{0}{deployment_id}\u{0}{container}")
-}
-
 /// Resolve a spec's requested replica count to the actual number of containers
 /// to run: default 1 (unset), clamped to `[1, MAX_REPLICAS]`. Shared by desired
 /// computation and health aggregation so both always agree on the replica count.
@@ -2280,778 +1731,9 @@ fn clamp_replicas(requested: Option<u32>) -> u32 {
     requested.unwrap_or(1).clamp(1, MAX_REPLICAS)
 }
 
-// ── Env merge helpers ─────────────────────────────────────────────────
-
-/// Merge env for one container in final precedence:
-/// base (plain + secret) → system env → per-container overrides. Later writes
-/// win on key conflict.
-/// Concise, human-readable detail for a probe `reqwest::Error`: prefer the
-/// underlying transport source (e.g. "connection refused", "timed out") over
-/// reqwest's verbose wrapper text, so the surfaced reason is actionable.
-fn probe_error_detail(e: &reqwest::Error) -> String {
-    use std::error::Error;
-    if let Some(src) = e.source() {
-        src.to_string()
-    } else if e.is_timeout() {
-        "timed out".to_string()
-    } else if e.is_connect() {
-        "connection error".to_string()
-    } else {
-        e.to_string()
-    }
-}
-
-/// Effective HTTP health-probe path for a container, honoring its public
-/// `health_check` spec (`rise.toml [containers.X.health_check]`):
-///
-/// - `health_check` **absent** (`None`) → `None`: HTTP probing is OPT-IN, so a
-///   container with no `health_check` declared is treated as ready as soon as
-///   it is *running* — matching Kubernetes, where a Pod with no readiness probe
-///   is Ready once the container is up.
-/// - `disabled = true` → `None`: the probe is explicitly turned off (the caller
-///   treats a *running* container as ready), matching the Kubernetes behavior
-///   where `health_check = false` removes the readiness/liveness probes.
-/// - `path` set → that path, normalized to a leading `/`.
-/// - set but no `path` (or an empty `path`) → the controller's default
-///   `health_path`.
-///
-/// Note: the reconcile loop probes once per tick, so the spec's fine-grained
-/// timing/threshold knobs (`period_seconds`, `failure_threshold`, …) and the
-/// separate `liveness_enabled`/`readiness_enabled` toggles are not applied on
-/// Docker — only `path` and `disabled` are. See the deployment-backends feature
-/// matrix.
-fn effective_health_path(
-    spec: &crate::server::deployment::models::ContainerSpec,
-    default_path: &str,
-) -> Option<String> {
-    // Probing is opt-in: no `health_check` declared → no probe (ready when
-    // running).
-    let hc = spec.health_check.as_ref()?;
-    if hc.disabled {
-        return None;
-    }
-    match hc.path.as_deref().filter(|p| !p.is_empty()) {
-        Some(p) if p.starts_with('/') => Some(p.to_string()),
-        Some(p) => Some(format!("/{p}")),
-        None => Some(default_path.to_string()),
-    }
-}
-
-/// Whether a deployment's containers should carry Traefik routing labels, by
-/// cutover strategy:
-///
-/// - `Recreate`: routable only when the deployment is the active one for its
-///   group (Rise-gated single cutover — today's behavior).
-/// - `HealthRolling`: ALWAYS routable. The old active and the new Deploying
-///   deployment both join the one group-scoped Traefik service immediately;
-///   Traefik's per-server health check drains the old servers as the new ones
-///   come UP, so there's no cutover gap (and no g1→g2 route-hash churn from a
-///   later `is_active` flip).
-///
-/// Pure (no `self`) so it can be unit-tested without a daemon.
-pub(crate) fn routable_for(
-    cutover_strategy: crate::server::settings::CutoverStrategy,
-    is_active: bool,
-) -> bool {
-    match cutover_strategy {
-        crate::server::settings::CutoverStrategy::Recreate => is_active,
-        crate::server::settings::CutoverStrategy::HealthRolling => true,
-    }
-}
-
-/// Group-scoped Traefik service name(s) a container spec's routes emit, used by
-/// the reconciler's `serverStatus` lookup so it queries the SAME service(s) the
-/// labels stamp. Mirrors [`container_builder::group_service_names`] but derives
-/// from the runtime [`ContainerSpec`] + [`RouteSpec`] set (what `reconcile_health`
-/// has on hand): a single-route container yields the bare base
-/// `{project}-{group}-{container}`; a multi-route container yields per-route
-/// `{base}-{idx}` names (longest path-prefix first, matching the renderer).
-///
-/// A port-less worker has no routes/service → empty. Routability is implicit:
-/// the caller only consults this in `HealthRolling` mode, where every deployment
-/// is routable.
-fn service_names_for_spec(
-    project: &str,
-    deployment_group: &str,
-    spec: &crate::server::deployment::models::ContainerSpec,
-    route_specs: &[crate::server::deployment::models::RouteSpec],
-) -> Vec<String> {
-    if spec.port.is_none() {
-        return Vec::new();
-    }
-    let base = container_builder::group_service_base(project, deployment_group, &spec.name);
-    // Routes for this container, sorted longest-path-prefix-first — the SAME
-    // ordering `render_traefik_labels_for` uses to index per-route services.
-    let mut routes: Vec<&crate::server::deployment::models::RouteSpec> = route_specs
-        .iter()
-        .filter(|r| r.container == spec.name)
-        .collect();
-    routes.sort_by_key(|r| std::cmp::Reverse(r.path.len()));
-    let route_count = routes.len();
-    if route_count == 0 {
-        return Vec::new();
-    }
-    (0..route_count)
-        .map(|idx| container_builder::group_service_name(&base, idx, route_count))
-        .collect()
-}
-
-/// Whether a single replica counts as READY, factored out of the per-replica
-/// loop in `reconcile_health` so the `serverStatus`/fallback selection is
-/// unit-testable without a daemon. Inputs:
-///
-/// - `rolling`: the group's cutover strategy is `HealthRolling`;
-/// - `has_health_path`: the container has an effective health path;
-/// - `running`: the live container is `running` on the daemon;
-/// - `api_available`: a Traefik API client is configured AND its serverStatus
-///   call for this container's service(s) succeeded;
-/// - `server_up`: when `api_available`, whether Traefik reports this container's
-///   server URL UP (`None` = absent from the map / no IP yet);
-/// - `probe_ok`: the result of Rise's own HTTP probe when one was performed
-///   (`Some(true/false)`), or `None` when no probe has been run yet.
-///
-/// Returns [`ReadyVerdict::NeedsProbe`] when the decision depends on Rise's own
-/// probe but `probe_ok` is `None` — the caller then runs `probe_container` and
-/// re-resolves (or passes the result back in). Keeps `reconcile_health`'s
-/// behavior identical: rolling uses `rolling_rotation_decision`; non-rolling
-/// probes when a health path exists, else gates on `running`.
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum ReadyVerdict {
-    Ready,
-    NotReady(String),
-    /// The verdict needs Rise's own HTTP probe; the caller must run it and fold
-    /// the result back in (via `probe_ok`).
-    NeedsProbe,
-}
-
-pub(crate) fn replica_ready(
-    rolling: bool,
-    has_health_path: bool,
-    running: bool,
-    api_available: bool,
-    server_up: Option<bool>,
-    probe_ok: Option<bool>,
-) -> ReadyVerdict {
-    if rolling {
-        match rolling_rotation_decision(has_health_path, running, api_available, server_up) {
-            RotationDecision::InRotation => ReadyVerdict::Ready,
-            RotationDecision::NotInRotation(reason) => ReadyVerdict::NotReady(reason),
-            RotationDecision::FallBackToProbe => match probe_ok {
-                Some(true) => ReadyVerdict::Ready,
-                Some(false) => ReadyVerdict::NotReady("health probe failed".to_string()),
-                None => ReadyVerdict::NeedsProbe,
-            },
-        }
-    } else if has_health_path {
-        // Non-rolling with a probe: readiness IS the probe result.
-        match probe_ok {
-            Some(true) => ReadyVerdict::Ready,
-            Some(false) => ReadyVerdict::NotReady("health probe failed".to_string()),
-            None => ReadyVerdict::NeedsProbe,
-        }
-    } else if running {
-        // No probe: a running container is ready (ready-when-running).
-        ReadyVerdict::Ready
-    } else {
-        ReadyVerdict::NotReady("not running (probe disabled)".to_string())
-    }
-}
-
-/// Outcome of the `HealthRolling` in-rotation decision for a single container.
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum RotationDecision {
-    /// The container's server is in Traefik's rotation (or running, for a
-    /// ready-when-running container) → counts toward deployment readiness.
-    InRotation,
-    /// The container is NOT in rotation; the string is a human-readable reason
-    /// for the deployment's not-ready diagnostic.
-    NotInRotation(String),
-    /// No authoritative Traefik signal is available (a `health_check` is
-    /// configured but the Traefik API is unset or errored). The caller must
-    /// MIRROR — fall back to Rise's own `probe_container` as the in-rotation
-    /// proxy (today's behavior).
-    FallBackToProbe,
-}
-
-/// Pure decision for whether a single container is "in rotation" in
-/// `HealthRolling` mode, given:
-///
-/// - `has_health_path`: the container has an effective health path (a
-///   `health_check` is configured AND not disabled);
-/// - `running`: the live container is `running` on the daemon;
-/// - `api_available`: a Traefik API client is configured AND its call for this
-///   service succeeded (so `server_up` is authoritative);
-/// - `server_up`: when `api_available`, whether Traefik's `serverStatus`
-///   reports this container's server URL (`http://{ip}:{port}`) as UP. `None`
-///   means the server URL was absent from the map (Traefik doesn't know it yet).
-///
-/// Rules (see the module / Step B spec):
-/// - has_health_path + api_available → in-rotation IFF Traefik reports the
-///   server UP; a DOWN/absent server is NOT in rotation (so the old deployment
-///   is kept until the new server is confirmed serving).
-/// - !has_health_path → ready-when-running: in-rotation IFF the container is
-///   `running` (Traefik routes to running servers immediately, no HC).
-/// - has_health_path + !api_available → `FallBackToProbe`: the caller mirrors
-///   with Rise's own probe (graceful degradation when no Traefik API).
-///
-/// Kept pure (no `self`, no I/O) so it can be unit-tested without a daemon.
-pub(crate) fn rolling_rotation_decision(
-    has_health_path: bool,
-    running: bool,
-    api_available: bool,
-    server_up: Option<bool>,
-) -> RotationDecision {
-    if !has_health_path {
-        // Ready-when-running: no health check, so Traefik routes to the server
-        // as soon as it's a running container.
-        return if running {
-            RotationDecision::InRotation
-        } else {
-            RotationDecision::NotInRotation("not running (no health check)".to_string())
-        };
-    }
-    // health_check configured.
-    if !api_available {
-        // No authoritative Traefik signal — mirror with Rise's own probe.
-        return RotationDecision::FallBackToProbe;
-    }
-    match server_up {
-        Some(true) => RotationDecision::InRotation,
-        Some(false) => RotationDecision::NotInRotation("Traefik reports server DOWN".to_string()),
-        None => RotationDecision::NotInRotation(
-            "Traefik does not yet report this server (absent from serverStatus)".to_string(),
-        ),
-    }
-}
-
-fn merge_container_env(
-    base_env: &[(String, String)],
-    system_env: &[(String, String)],
-    injected_hosts: &[(String, String)],
-    spec: &crate::server::deployment::models::ContainerSpec,
-    env_name: Option<&str>,
-) -> Vec<(String, String)> {
-    let mut env: Vec<(String, String)> = base_env.to_vec();
-    for (k, v) in system_env {
-        upsert_env(&mut env, k, v);
-    }
-    // Cross-container service-discovery hosts (`RISE_CONTAINER_HOST__<NAME>`).
-    // Only added when not already present, so user globals win; per-container
-    // `env_overrides` below run last and can shadow them too — matching the K8s
-    // precedence in `webhook.rs`.
-    for (k, v) in injected_hosts {
-        if !env.iter().any(|(ek, _)| ek == k) {
-            env.push((k.clone(), v.clone()));
-        }
-    }
-    for over in &spec.env_overrides {
-        // Per-container secret overrides are rejected at request time.
-        if over.is_secret {
-            continue;
-        }
-        if let Some(ref target_env) = over.for_environment {
-            if env_name != Some(target_env.as_str()) {
-                continue;
-            }
-        }
-        upsert_env(&mut env, &over.key, &over.value);
-    }
-    env
-}
-
-fn upsert_env(env: &mut Vec<(String, String)>, key: &str, value: &str) {
-    if let Some(existing) = env.iter_mut().find(|(k, _)| k == key) {
-        existing.1 = value.to_string();
-    } else {
-        env.push((key.to_string(), value.to_string()));
-    }
-}
-
-/// Stable sha256 of a merged env vector, used as the drift label. Hashes the
-/// *entire* set (plain + system/RISE_* + secret) over a deterministically
-/// key-sorted copy with length-prefixed key/value framing, so reordering can't
-/// change the digest while any add/edit/delete of any variable does. Editing or
-/// deleting any env var therefore changes the `env-hash` label and forces the
-/// reconciler to recreate the container.
-pub fn hash_env(env: &[(String, String)]) -> String {
-    let mut sorted: Vec<&(String, String)> = env.iter().collect();
-    sorted.sort_by(|a, b| a.0.cmp(&b.0).then_with(|| a.1.cmp(&b.1)));
-    let mut hasher = Sha256::new();
-    for (k, v) in sorted {
-        hasher.update((k.len() as u64).to_le_bytes());
-        hasher.update(k.as_bytes());
-        hasher.update((v.len() as u64).to_le_bytes());
-        hasher.update(v.as_bytes());
-    }
-    hasher
-        .finalize()
-        .iter()
-        .map(|b| format!("{b:02x}"))
-        .collect()
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::server::deployment::controller::docker::container_builder::DesiredRoute;
-
-    fn desired(container: &str, image: &str, hash: &str) -> DesiredContainer {
-        DesiredContainer {
-            project: "myapp".to_string(),
-            access_class: "public".to_string(),
-            deployment_group: "default".to_string(),
-            deployment_id: "20260101-120000".to_string(),
-            deployment_uuid: "11111111-1111-1111-1111-111111111111".to_string(),
-            container: container.to_string(),
-            environment: None,
-            image: image.to_string(),
-            port: Some(8080),
-            cpu: "500m".to_string(),
-            memory: "256Mi".to_string(),
-            env: vec![],
-            env_hash: hash.to_string(),
-            routes: vec![DesiredRoute {
-                hosts: vec!["myapp.rise.dev".to_string()],
-                path_prefix: None,
-            }],
-            routable: true,
-            // Fixed sentinel route-hash for diff tests; the reconciler computes
-            // the real value via `route_hash_for`. Tests that exercise routing
-            // drift override this and the matching actual label.
-            route_hash: "rh-active".to_string(),
-            // Seed generation (the diff resolves the real value before apply).
-            generation: 1,
-            replica: 0,
-            health_path: Some("/".to_string()),
-            health_check_interval_secs: None,
-            health_check_timeout_secs: None,
-        }
-    }
-
-    /// Stable identity key for the `desired()` helper's slot, used to build the
-    /// `identity` field of expected Create/Recreate actions.
-    fn identity_of(d: &DesiredContainer) -> String {
-        identity_key(
-            &d.project,
-            &d.deployment_group,
-            &d.deployment_id,
-            &d.container,
-            d.replica,
-        )
-    }
-
-    /// Empty protected-deployment-ids set for the common case where no
-    /// deployment failed desired computation.
-    fn no_protected() -> HashSet<String> {
-        HashSet::new()
-    }
-
-    /// The resolved generation-ful name for the `desired()` slot at a given
-    /// generation (the name a Create/Recreate action carries).
-    fn name_of_gen(d: &DesiredContainer, generation: u32) -> String {
-        container_builder::container_name(
-            "rise",
-            &d.project,
-            &d.deployment_group,
-            &d.deployment_id,
-            &d.container,
-            d.replica,
-            generation,
-        )
-    }
-
-    /// A live container belonging to the `desired()` helper's deployment, at a
-    /// given generation. Its name carries the `_g{generation}` suffix and it
-    /// carries the full identity-label set so the diff can match it.
-    fn actual_for_gen(
-        d: &DesiredContainer,
-        image: &str,
-        env_hash: &str,
-        generation: u32,
-    ) -> ActualContainer {
-        ActualContainer {
-            id: "cid".to_string(),
-            name: name_of_gen(d, generation),
-            project: Some(d.project.clone()),
-            deployment_group: Some(d.deployment_group.clone()),
-            container: Some(d.container.clone()),
-            deployment_id_label: Some(d.deployment_id.clone()),
-            generation,
-            replica: d.replica,
-            image_label: Some(image.to_string()),
-            env_hash_label: Some(env_hash.to_string()),
-            route_hash_label: Some("rh-active".to_string()),
-            state: Some("running".to_string()),
-        }
-    }
-
-    /// A live container at generation 1 (the common case for most diff tests).
-    fn actual_for(d: &DesiredContainer, image: &str, env_hash: &str) -> ActualContainer {
-        actual_for_gen(d, image, env_hash, 1)
-    }
-
-    #[test]
-    fn diff_creates_missing() {
-        let d = desired("app", "img:1", "h1");
-        let actions =
-            diff_desired_vs_actual(std::slice::from_ref(&d), &[], "rise", &no_protected());
-        assert_eq!(
-            actions,
-            vec![ReconcileAction::Create {
-                identity: identity_of(&d),
-                name: name_of_gen(&d, 1),
-                generation: 1,
-            }]
-        );
-    }
-
-    #[test]
-    fn diff_no_action_when_matched() {
-        let d = desired("app", "img:1", "h1");
-        let actual = vec![actual_for(&d, "img:1", "h1")];
-        let actions = diff_desired_vs_actual(&[d], &actual, "rise", &no_protected());
-        assert!(actions.is_empty());
-    }
-
-    #[test]
-    fn diff_recreates_on_image_drift() {
-        let d = desired("app", "img:2", "h1");
-        let actual = vec![actual_for(&d, "img:1", "h1")];
-        let actions =
-            diff_desired_vs_actual(std::slice::from_ref(&d), &actual, "rise", &no_protected());
-        assert_eq!(
-            actions,
-            vec![ReconcileAction::Recreate {
-                identity: identity_of(&d),
-                name: name_of_gen(&d, 2),
-                existing_id: "cid".to_string(),
-                generation: 2,
-            }]
-        );
-    }
-
-    #[test]
-    fn diff_recreates_on_env_hash_drift() {
-        let d = desired("app", "img:1", "h2");
-        let actual = vec![actual_for(&d, "img:1", "h1")];
-        let actions =
-            diff_desired_vs_actual(std::slice::from_ref(&d), &actual, "rise", &no_protected());
-        assert_eq!(
-            actions,
-            vec![ReconcileAction::Recreate {
-                identity: identity_of(&d),
-                name: name_of_gen(&d, 2),
-                existing_id: "cid".to_string(),
-                generation: 2,
-            }]
-        );
-    }
-
-    #[test]
-    fn diff_recreates_when_matched_but_not_running() {
-        // Image + env match, but the container exited / never started — it must
-        // be recreated so the deployment recovers.
-        let d = desired("app", "img:1", "h1");
-        for state in [Some("exited"), Some("created"), Some("dead"), None] {
-            let actual = vec![ActualContainer {
-                state: state.map(str::to_string),
-                ..actual_for(&d, "img:1", "h1")
-            }];
-            let actions =
-                diff_desired_vs_actual(std::slice::from_ref(&d), &actual, "rise", &no_protected());
-            assert_eq!(
-                actions,
-                vec![ReconcileAction::Recreate {
-                    identity: identity_of(&d),
-                    name: name_of_gen(&d, 2),
-                    existing_id: "cid".to_string(),
-                    generation: 2,
-                }],
-                "state {state:?} should force recreate"
-            );
-        }
-    }
-
-    #[test]
-    fn diff_leaves_transient_states_to_docker() {
-        // Image + env match and the container is in a daemon-managed transient
-        // state: `running` (healthy), `restarting` (Docker's restart policy
-        // applying backoff after a crash) or `paused` (operator-initiated).
-        // None of these may be force-recreated — doing so each tick would reset
-        // Docker's restart backoff. The health probe drives Unhealthy status
-        // independently.
-        let d = desired("app", "img:1", "h1");
-        for state in [Some("running"), Some("restarting"), Some("paused")] {
-            let actual = vec![ActualContainer {
-                state: state.map(str::to_string),
-                ..actual_for(&d, "img:1", "h1")
-            }];
-            let actions =
-                diff_desired_vs_actual(std::slice::from_ref(&d), &actual, "rise", &no_protected());
-            assert!(
-                actions.is_empty(),
-                "state {state:?} must not force a recreate"
-            );
-        }
-    }
-
-    #[test]
-    fn diff_removes_orphan() {
-        let actual = vec![ActualContainer {
-            id: "old".to_string(),
-            name: "rise_myapp_default_oldid_app_g1".to_string(),
-            project: Some("myapp".to_string()),
-            deployment_group: Some("default".to_string()),
-            container: Some("app".to_string()),
-            deployment_id_label: Some("oldid".to_string()),
-            generation: 1,
-            replica: 0,
-            image_label: Some("img:1".to_string()),
-            env_hash_label: Some("h1".to_string()),
-            route_hash_label: Some("rh-active".to_string()),
-            state: Some("running".to_string()),
-        }];
-        let actions = diff_desired_vs_actual(&[], &actual, "rise", &no_protected());
-        assert_eq!(
-            actions,
-            vec![ReconcileAction::Remove {
-                id: "old".to_string(),
-                name: "rise_myapp_default_oldid_app_g1".to_string()
-            }]
-        );
-    }
-
-    #[test]
-    fn diff_supersession_creates_new_removes_old() {
-        // New deployment desired; old deployment's container still present.
-        let new = desired("app", "img:2", "h2");
-        let actual = vec![ActualContainer {
-            id: "old".to_string(),
-            name: "rise_myapp_default_oldid_app_g1".to_string(),
-            project: Some("myapp".to_string()),
-            deployment_group: Some("default".to_string()),
-            container: Some("app".to_string()),
-            deployment_id_label: Some("oldid".to_string()),
-            generation: 1,
-            replica: 0,
-            image_label: Some("img:1".to_string()),
-            env_hash_label: Some("h1".to_string()),
-            route_hash_label: Some("rh-active".to_string()),
-            state: Some("running".to_string()),
-        }];
-        let actions =
-            diff_desired_vs_actual(std::slice::from_ref(&new), &actual, "rise", &no_protected());
-        assert_eq!(
-            actions,
-            vec![
-                ReconcileAction::Create {
-                    identity: identity_of(&new),
-                    name: name_of_gen(&new, 1),
-                    generation: 1,
-                },
-                ReconcileAction::Remove {
-                    id: "old".to_string(),
-                    name: "rise_myapp_default_oldid_app_g1".to_string()
-                },
-            ]
-        );
-    }
-
-    #[test]
-    fn diff_protects_container_when_its_deployment_failed_to_compute() {
-        // A deployment whose desired computation failed this tick (e.g. a brief
-        // KMS/DB hiccup) is absent from `desired`, so its still-running container
-        // looks orphaned. With its deployment-id protected, the diff must NOT
-        // remove it — matching the K8s model (children untouched on failed sync).
-        let actual = vec![ActualContainer {
-            id: "live".to_string(),
-            name: "rise_myapp_default_protectedid_app_g1".to_string(),
-            project: Some("myapp".to_string()),
-            deployment_group: Some("default".to_string()),
-            container: Some("app".to_string()),
-            deployment_id_label: Some("protectedid".to_string()),
-            generation: 1,
-            replica: 0,
-            image_label: Some("img:1".to_string()),
-            env_hash_label: Some("h1".to_string()),
-            route_hash_label: Some("rh-active".to_string()),
-            state: Some("running".to_string()),
-        }];
-        let mut protected = HashSet::new();
-        protected.insert("protectedid".to_string());
-        let actions = diff_desired_vs_actual(&[], &actual, "rise", &protected);
-        assert!(
-            actions.is_empty(),
-            "protected deployment's container must not be GC'd"
-        );
-    }
-
-    #[test]
-    fn diff_still_gcs_unprotected_orphan_alongside_protected() {
-        // Only the failed deployment's container is protected; a genuinely
-        // orphaned container belonging to a different, successfully-computed
-        // deployment is still removed.
-        let actual = vec![
-            ActualContainer {
-                id: "protected".to_string(),
-                name: "rise_myapp_default_protectedid_app_g1".to_string(),
-                project: Some("myapp".to_string()),
-                deployment_group: Some("default".to_string()),
-                container: Some("app".to_string()),
-                deployment_id_label: Some("protectedid".to_string()),
-                generation: 1,
-                replica: 0,
-                image_label: Some("img:1".to_string()),
-                env_hash_label: Some("h1".to_string()),
-                route_hash_label: Some("rh-active".to_string()),
-                state: Some("running".to_string()),
-            },
-            ActualContainer {
-                id: "orphan".to_string(),
-                name: "rise_myapp_default_orphanid_app_g1".to_string(),
-                project: Some("myapp".to_string()),
-                deployment_group: Some("default".to_string()),
-                container: Some("app".to_string()),
-                deployment_id_label: Some("orphanid".to_string()),
-                generation: 1,
-                replica: 0,
-                image_label: Some("img:1".to_string()),
-                env_hash_label: Some("h1".to_string()),
-                route_hash_label: Some("rh-active".to_string()),
-                state: Some("running".to_string()),
-            },
-        ];
-        let mut protected = HashSet::new();
-        protected.insert("protectedid".to_string());
-        let actions = diff_desired_vs_actual(&[], &actual, "rise", &protected);
-        assert_eq!(
-            actions,
-            vec![ReconcileAction::Remove {
-                id: "orphan".to_string(),
-                name: "rise_myapp_default_orphanid_app_g1".to_string()
-            }],
-            "only the unprotected orphan should be removed"
-        );
-    }
-
-    #[test]
-    fn diff_recreates_on_route_hash_drift_active_gains_labels() {
-        // A deployment that just became active: desired now carries the active
-        // route-hash, but the live container was created while non-routable and
-        // still bears the empty/non-routable hash. Image + env match, but the
-        // routing changed — Docker can't edit labels in place, so the container
-        // must be recreated WITH the Traefik labels.
-        let d = desired("app", "img:1", "h1"); // route_hash = "rh-active"
-        let actual = vec![ActualContainer {
-            // Was created non-routable (empty route-hash).
-            route_hash_label: Some(String::new()),
-            ..actual_for(&d, "img:1", "h1")
-        }];
-        let actions =
-            diff_desired_vs_actual(std::slice::from_ref(&d), &actual, "rise", &no_protected());
-        assert_eq!(
-            actions,
-            vec![ReconcileAction::Recreate {
-                identity: identity_of(&d),
-                name: name_of_gen(&d, 2),
-                existing_id: "cid".to_string(),
-                generation: 2,
-            }]
-        );
-    }
-
-    #[test]
-    fn diff_recreates_on_route_hash_drift_deactivated_loses_labels() {
-        // A deployment that stopped being active: desired is now non-routable
-        // (empty route-hash) while the live container still carries the active
-        // routing hash. Image + env match, but its Traefik labels must be
-        // removed — force a recreate.
-        let mut d = desired("app", "img:1", "h1");
-        d.routable = false;
-        d.route_hash = String::new();
-        let actual = vec![ActualContainer {
-            route_hash_label: Some("rh-active".to_string()),
-            ..actual_for(&d, "img:1", "h1")
-        }];
-        let actions =
-            diff_desired_vs_actual(std::slice::from_ref(&d), &actual, "rise", &no_protected());
-        assert_eq!(
-            actions,
-            vec![ReconcileAction::Recreate {
-                identity: identity_of(&d),
-                name: name_of_gen(&d, 2),
-                existing_id: "cid".to_string(),
-                generation: 2,
-            }]
-        );
-    }
-
-    #[test]
-    fn diff_no_action_when_route_image_env_all_match() {
-        // Idempotence: once image + env + route-hash all match and the container
-        // is running, the diff emits NO action — no recreate oscillation.
-        let d = desired("app", "img:1", "h1"); // route_hash = "rh-active"
-        let actual = vec![actual_for(&d, "img:1", "h1")];
-        let actions =
-            diff_desired_vs_actual(std::slice::from_ref(&d), &actual, "rise", &no_protected());
-        assert!(actions.is_empty(), "matching container must not recreate");
-    }
-
-    #[test]
-    fn diff_no_action_when_matched_with_different_generation_suffix() {
-        // The KEY anti-infinite-recreate test: a stable container (no image/env/
-        // route drift, running) that was recreated several times — so its live
-        // name carries `_g5` and its generation label is 5 — must still match the
-        // desired slot by IDENTITY and produce ZERO actions. The generation never
-        // feeds any matching key or hash, so the `_g{n}` suffix is irrelevant.
-        let d = desired("app", "img:1", "h1"); // route_hash = "rh-active"
-        let actual = vec![actual_for_gen(&d, "img:1", "h1", 5)];
-        assert!(
-            actual[0].name.ends_with("_g5"),
-            "fixture must carry a non-1 generation suffix"
-        );
-        let actions =
-            diff_desired_vs_actual(std::slice::from_ref(&d), &actual, "rise", &no_protected());
-        assert!(
-            actions.is_empty(),
-            "a stable container must not recreate regardless of its _g{{n}} suffix"
-        );
-    }
-
-    #[test]
-    fn diff_recreate_bumps_generation_from_actual() {
-        // A drifted match at generation 5 recreates at generation 6 (actual + 1),
-        // and the carried name reflects the bumped generation.
-        let d = desired("app", "img:2", "h1"); // image drift
-        let actual = vec![actual_for_gen(&d, "img:1", "h1", 5)];
-        let actions =
-            diff_desired_vs_actual(std::slice::from_ref(&d), &actual, "rise", &no_protected());
-        assert_eq!(
-            actions,
-            vec![ReconcileAction::Recreate {
-                identity: identity_of(&d),
-                name: name_of_gen(&d, 6),
-                existing_id: "cid".to_string(),
-                generation: 6,
-            }]
-        );
-    }
-
-    #[test]
-    fn diff_create_uses_generation_one() {
-        // A missing slot always creates at generation 1.
-        let d = desired("app", "img:1", "h1");
-        let actions =
-            diff_desired_vs_actual(std::slice::from_ref(&d), &[], "rise", &no_protected());
-        assert_eq!(
-            actions,
-            vec![ReconcileAction::Create {
-                identity: identity_of(&d),
-                name: name_of_gen(&d, 1),
-                generation: 1,
-            }]
-        );
-    }
 
     #[test]
     fn controller_class_matching() {
@@ -3072,602 +1754,6 @@ mod tests {
     }
 
     #[test]
-    fn merge_env_precedence() {
-        use crate::server::deployment::models::{ContainerSpec, EnvOverride};
-        let base = vec![("FOO".to_string(), "base".to_string())];
-        let system = vec![
-            ("FOO".to_string(), "system".to_string()),
-            ("RISE_APP_URL".to_string(), "url".to_string()),
-        ];
-        let spec = ContainerSpec {
-            name: "app".to_string(),
-            image: None,
-            port: Some(8080),
-            replicas: None,
-            cpu: None,
-            memory: None,
-            env_overrides: vec![EnvOverride {
-                key: "FOO".to_string(),
-                value: "override".to_string(),
-                is_secret: false,
-                is_protected: None,
-                source: None,
-                for_environment: None,
-            }],
-            health_check: None,
-        };
-        let merged = merge_container_env(&base, &system, &[], &spec, None);
-        let foo = merged.iter().find(|(k, _)| k == "FOO").unwrap();
-        assert_eq!(foo.1, "override");
-        assert!(merged
-            .iter()
-            .any(|(k, v)| k == "RISE_APP_URL" && v == "url"));
-    }
-
-    #[test]
-    fn effective_health_path_honors_spec() {
-        use crate::server::deployment::models::{ContainerSpec, HealthCheckSpec};
-        let base = ContainerSpec {
-            name: "app".to_string(),
-            image: None,
-            port: Some(8080),
-            replicas: None,
-            cpu: None,
-            memory: None,
-            env_overrides: vec![],
-            health_check: None,
-        };
-        // No spec → None: probing is opt-in, so an absent `health_check` means
-        // no HTTP probe (the container is ready when running).
-        assert_eq!(
-            effective_health_path(&base, "/"),
-            None,
-            "absent health_check → no probe"
-        );
-        // `health_check` set but no path → controller default.
-        let default_path = ContainerSpec {
-            health_check: Some(HealthCheckSpec::default()),
-            ..base.clone()
-        };
-        assert_eq!(
-            effective_health_path(&default_path, "/livez").as_deref(),
-            Some("/livez"),
-            "set-but-no-path → controller default"
-        );
-        // disabled → None (probe turned off).
-        let disabled = ContainerSpec {
-            health_check: Some(HealthCheckSpec {
-                disabled: true,
-                ..Default::default()
-            }),
-            ..base.clone()
-        };
-        assert_eq!(effective_health_path(&disabled, "/"), None, "disabled");
-        // explicit path used verbatim.
-        let with_path = ContainerSpec {
-            health_check: Some(HealthCheckSpec {
-                path: Some("/healthz".to_string()),
-                ..Default::default()
-            }),
-            ..base.clone()
-        };
-        assert_eq!(
-            effective_health_path(&with_path, "/").as_deref(),
-            Some("/healthz")
-        );
-        // path missing a leading slash is normalized.
-        let no_slash = ContainerSpec {
-            health_check: Some(HealthCheckSpec {
-                path: Some("healthz".to_string()),
-                ..Default::default()
-            }),
-            ..base.clone()
-        };
-        assert_eq!(
-            effective_health_path(&no_slash, "/").as_deref(),
-            Some("/healthz")
-        );
-        // empty path falls back to the default.
-        let empty = ContainerSpec {
-            health_check: Some(HealthCheckSpec {
-                path: Some(String::new()),
-                ..Default::default()
-            }),
-            ..base.clone()
-        };
-        assert_eq!(
-            effective_health_path(&empty, "/livez").as_deref(),
-            Some("/livez")
-        );
-    }
-
-    #[test]
-    fn merge_env_injects_container_hosts_unless_user_set() {
-        use crate::server::deployment::models::{ContainerSpec, EnvOverride};
-        // A user global and a per-container override both targeting the same
-        // discovery key must win over the auto-injected sibling host.
-        let base = vec![(
-            "RISE_CONTAINER_HOST__API".to_string(),
-            "user-global:1".to_string(),
-        )];
-        // Injected hosts use the GROUP-scoped, deployment-id-FREE alias
-        // (`{prefix}_{project}_{group}_{container}`).
-        let injected = vec![
-            (
-                "RISE_CONTAINER_HOST__API".to_string(),
-                "rise_myapp_default_api:8080".to_string(),
-            ),
-            (
-                "RISE_CONTAINER_HOST__REDIS".to_string(),
-                "rise_myapp_default_redis:6379".to_string(),
-            ),
-        ];
-        let spec = ContainerSpec {
-            name: "web".to_string(),
-            image: None,
-            port: Some(8080),
-            replicas: None,
-            cpu: None,
-            memory: None,
-            env_overrides: vec![EnvOverride {
-                key: "RISE_CONTAINER_HOST__REDIS".to_string(),
-                value: "override-redis:6379".to_string(),
-                is_secret: false,
-                is_protected: None,
-                source: None,
-                for_environment: None,
-            }],
-            health_check: None,
-        };
-        let merged = merge_container_env(&base, &[], &injected, &spec, None);
-        // User global wins; injected value is not appended a second time.
-        let api: Vec<_> = merged
-            .iter()
-            .filter(|(k, _)| k == "RISE_CONTAINER_HOST__API")
-            .collect();
-        assert_eq!(api.len(), 1);
-        assert_eq!(api[0].1, "user-global:1");
-        // Per-container override shadows the injected sibling host.
-        let redis = merged
-            .iter()
-            .find(|(k, _)| k == "RISE_CONTAINER_HOST__REDIS")
-            .unwrap();
-        assert_eq!(redis.1, "override-redis:6379");
-    }
-
-    #[test]
-    fn merge_env_skips_non_matching_environment_override() {
-        use crate::server::deployment::models::{ContainerSpec, EnvOverride};
-        let spec = ContainerSpec {
-            name: "app".to_string(),
-            image: None,
-            port: None,
-            replicas: None,
-            cpu: None,
-            memory: None,
-            env_overrides: vec![EnvOverride {
-                key: "ONLY_PROD".to_string(),
-                value: "1".to_string(),
-                is_secret: false,
-                is_protected: None,
-                source: None,
-                for_environment: Some("production".to_string()),
-            }],
-            health_check: None,
-        };
-        let merged = merge_container_env(&[], &[], &[], &spec, Some("staging"));
-        assert!(!merged.iter().any(|(k, _)| k == "ONLY_PROD"));
-        let merged_prod = merge_container_env(&[], &[], &[], &spec, Some("production"));
-        assert!(merged_prod.iter().any(|(k, _)| k == "ONLY_PROD"));
-    }
-
-    #[test]
-    fn hash_env_is_order_independent_but_value_sensitive() {
-        let a = vec![
-            ("A".to_string(), "1".to_string()),
-            ("B".to_string(), "2".to_string()),
-        ];
-        let b = vec![
-            ("B".to_string(), "2".to_string()),
-            ("A".to_string(), "1".to_string()),
-        ];
-        // Reordering the same set yields the same hash.
-        assert_eq!(hash_env(&a), hash_env(&b));
-        // Changing a value changes the hash.
-        let changed = vec![
-            ("A".to_string(), "1".to_string()),
-            ("B".to_string(), "3".to_string()),
-        ];
-        assert_ne!(hash_env(&a), hash_env(&changed));
-        // Deleting a var changes the hash.
-        let deleted = vec![("A".to_string(), "1".to_string())];
-        assert_ne!(hash_env(&a), hash_env(&deleted));
-        // Adding a plain var changes the hash (the core drift bug this fixes).
-        let mut added = a.clone();
-        added.push(("C".to_string(), "3".to_string()));
-        assert_ne!(hash_env(&a), hash_env(&added));
-    }
-
-    #[test]
-    fn hash_env_avoids_delimiter_collisions() {
-        // Length-prefixed framing means `{A:"B", : "C"}`-style splits can't
-        // collide with `{A:"BC"}`-style merges.
-        let split = vec![
-            ("A".to_string(), "B".to_string()),
-            ("C".to_string(), "D".to_string()),
-        ];
-        let merged = vec![("A".to_string(), "BCD".to_string())];
-        assert_ne!(hash_env(&split), hash_env(&merged));
-    }
-
-    fn inspected_running() -> InspectedContainer {
-        InspectedContainer {
-            status: Some("running".to_string()),
-            running: true,
-            started_at: Some("2026-01-01T12:00:00Z".to_string()),
-            finished_at: None,
-            exit_code: None,
-            restart_count: Some(0),
-            health: None,
-            error: None,
-            ip: Some("172.20.0.5".to_string()),
-            published_host_port: Some("49153".to_string()),
-        }
-    }
-
-    fn inspected_exited(code: i64) -> InspectedContainer {
-        InspectedContainer {
-            status: Some("exited".to_string()),
-            running: false,
-            started_at: Some("2026-01-01T12:00:00Z".to_string()),
-            finished_at: Some("2026-01-01T12:00:05Z".to_string()),
-            exit_code: Some(code),
-            restart_count: Some(2),
-            health: None,
-            error: None,
-            ip: None,
-            published_host_port: None,
-        }
-    }
-
-    #[test]
-    fn build_pod_status_running_maps_running_phase_and_state() {
-        let named = vec![("app".to_string(), Some(inspected_running()))];
-        let ps = build_pod_status(&named, true);
-
-        assert_eq!(ps["desired_replicas"], 1);
-        assert_eq!(ps["current_replicas"], 1);
-        assert_eq!(ps["ready_replicas"], 1);
-
-        let pod = &ps["pods"][0];
-        assert_eq!(pod["name"], "app");
-        assert_eq!(pod["phase"], "Running");
-        assert_eq!(pod["conditions"], serde_json::json!([]));
-
-        let c = &pod["containers"][0];
-        assert_eq!(c["name"], "app");
-        assert_eq!(c["ready"], true);
-        assert_eq!(c["restart_count"], 0);
-        assert_eq!(c["state"]["state_type"], "running");
-        assert_eq!(c["state"]["started_at"], "2026-01-01T12:00:00Z");
-    }
-
-    #[test]
-    fn build_pod_status_exited_maps_failed_phase_and_terminated_state() {
-        let named = vec![("app".to_string(), Some(inspected_exited(1)))];
-        // Even if the overall verdict were true, a non-running container is not
-        // counted ready; here the verdict is false anyway.
-        let ps = build_pod_status(&named, false);
-
-        assert_eq!(ps["desired_replicas"], 1);
-        assert_eq!(ps["current_replicas"], 0);
-        assert_eq!(ps["ready_replicas"], 0);
-
-        let pod = &ps["pods"][0];
-        assert_eq!(pod["phase"], "Failed");
-
-        let c = &pod["containers"][0];
-        assert_eq!(c["ready"], false);
-        assert_eq!(c["restart_count"], 2);
-        assert_eq!(c["state"]["state_type"], "terminated");
-        assert_eq!(c["state"]["exit_code"], 1);
-        assert_eq!(c["state"]["finished_at"], "2026-01-01T12:00:05Z");
-        // Non-zero exit synthesizes a reason when the daemon gave none.
-        assert_eq!(c["state"]["reason"], "Error (exit 1)");
-    }
-
-    #[test]
-    fn build_pod_status_missing_inspection_is_waiting_unknown() {
-        // A container with no inspection (not yet created) → Unknown phase,
-        // waiting state, not running, not ready.
-        let named = vec![("app".to_string(), None)];
-        let ps = build_pod_status(&named, true);
-        assert_eq!(ps["current_replicas"], 0);
-        assert_eq!(ps["ready_replicas"], 0);
-        let pod = &ps["pods"][0];
-        assert_eq!(pod["phase"], "Unknown");
-        assert_eq!(pod["containers"][0]["state"]["state_type"], "waiting");
-        assert_eq!(pod["containers"][0]["state"]["reason"], "ContainerCreating");
-    }
-
-    #[test]
-    fn build_pod_status_running_but_not_ready_overall_is_not_container_ready() {
-        // The container is running but the deployment verdict is false (another
-        // container failing): this container is current but not ready.
-        let named = vec![("app".to_string(), Some(inspected_running()))];
-        let ps = build_pod_status(&named, false);
-        assert_eq!(ps["current_replicas"], 1);
-        assert_eq!(ps["ready_replicas"], 0);
-        assert_eq!(ps["pods"][0]["containers"][0]["ready"], false);
-    }
-
-    // ── Replicas: diff scale up/down + rolling-recreate throttle ───────────
-
-    /// Build the `desired()` slot for a specific replica index.
-    fn desired_replica(replica: u32) -> DesiredContainer {
-        let mut d = desired("app", "img:1", "h1");
-        d.replica = replica;
-        d
-    }
-
-    /// A live (matched) container for the given replica/state/image. Carries the
-    /// full identity-label set including the replica so the diff matches it.
-    fn actual_replica(replica: u32, state: &str, image: &str) -> ActualContainer {
-        let d = desired_replica(replica);
-        ActualContainer {
-            id: format!("cid-r{replica}"),
-            state: Some(state.to_string()),
-            ..actual_for(&d, image, "h1")
-        }
-    }
-
-    #[test]
-    fn diff_scale_up_creates_new_replica_indices() {
-        // Desired 3 replicas (r0..r2); only r0 exists → r1 and r2 are Created,
-        // r0 is left untouched (image/env/route all match).
-        let desired_all = vec![desired_replica(0), desired_replica(1), desired_replica(2)];
-        let actual = vec![actual_replica(0, "running", "img:1")];
-        let actions = diff_desired_vs_actual(&desired_all, &actual, "rise", &no_protected());
-        let creates: Vec<u32> = actions
-            .iter()
-            .filter_map(|a| match a {
-                ReconcileAction::Create { name, .. } if name.contains("_r1_") => Some(1),
-                ReconcileAction::Create { name, .. } if name.contains("_r2_") => Some(2),
-                _ => None,
-            })
-            .collect();
-        assert_eq!(
-            creates,
-            vec![1, 2],
-            "scale-up creates the new replica slots"
-        );
-        assert!(
-            !actions.iter().any(|a| matches!(
-                a,
-                ReconcileAction::Recreate { .. } | ReconcileAction::Remove { .. }
-            )),
-            "scale-up must not recreate or remove the existing replica"
-        );
-    }
-
-    #[test]
-    fn diff_scale_down_removes_surplus_replicas() {
-        // Desired 1 replica (r0); r0..r2 exist → r1 and r2 are surplus → Removed.
-        let desired_all = vec![desired_replica(0)];
-        let actual = vec![
-            actual_replica(0, "running", "img:1"),
-            actual_replica(1, "running", "img:1"),
-            actual_replica(2, "running", "img:1"),
-        ];
-        let actions = diff_desired_vs_actual(&desired_all, &actual, "rise", &no_protected());
-        let mut removed: Vec<String> = actions
-            .iter()
-            .filter_map(|a| match a {
-                ReconcileAction::Remove { id, .. } => Some(id.clone()),
-                _ => None,
-            })
-            .collect();
-        removed.sort();
-        assert_eq!(removed, vec!["cid-r1".to_string(), "cid-r2".to_string()]);
-        assert!(
-            !actions
-                .iter()
-                .any(|a| matches!(a, ReconcileAction::Create { .. })),
-            "scale-down adds no containers"
-        );
-    }
-
-    #[test]
-    fn diff_scale_down_respects_protected_deployment_ids() {
-        // A surplus replica whose deployment-id is protected is NOT removed.
-        let desired_all = vec![desired_replica(0)];
-        let actual = vec![
-            actual_replica(0, "running", "img:1"),
-            actual_replica(1, "running", "img:1"),
-        ];
-        let mut protected = HashSet::new();
-        protected.insert("20260101-120000".to_string()); // desired()'s deployment_id
-        let actions = diff_desired_vs_actual(&desired_all, &actual, "rise", &protected);
-        assert!(
-            !actions
-                .iter()
-                .any(|a| matches!(a, ReconcileAction::Remove { .. })),
-            "protected surplus replica must not be GC'd"
-        );
-    }
-
-    /// Health map marking every given identity healthy.
-    fn all_healthy(actual: &[ActualContainer]) -> HashMap<String, bool> {
-        actual
-            .iter()
-            .filter_map(|a| a.identity().map(|id| (id, true)))
-            .collect()
-    }
-
-    #[test]
-    fn rolling_throttle_running_drifted_emits_exactly_one_recreate() {
-        // INVARIANT (a): 3 running + drifted (image) + healthy replicas →
-        // the diff produces 3 Recreates, but the rolling throttle admits exactly
-        // ONE (the lowest replica index, r0).
-        let desired_all = vec![desired_replica(0), desired_replica(1), desired_replica(2)];
-        let actual = vec![
-            actual_replica(0, "running", "img:OLD"),
-            actual_replica(1, "running", "img:OLD"),
-            actual_replica(2, "running", "img:OLD"),
-        ];
-        let raw = diff_desired_vs_actual(&desired_all, &actual, "rise", &no_protected());
-        assert_eq!(
-            raw.iter()
-                .filter(|a| matches!(a, ReconcileAction::Recreate { .. }))
-                .count(),
-            3,
-            "diff alone wants to recreate all three"
-        );
-        let healthy = all_healthy(&actual);
-        let throttled = filter_rolling_actions(raw, &actual, &healthy);
-        let recreates: Vec<&ReconcileAction> = throttled
-            .iter()
-            .filter(|a| matches!(a, ReconcileAction::Recreate { .. }))
-            .collect();
-        assert_eq!(recreates.len(), 1, "rolling admits exactly one per spec");
-        match recreates[0] {
-            ReconcileAction::Recreate { name, .. } => {
-                assert!(
-                    name.contains("_r0_"),
-                    "lowest replica index rolls first: {name}"
-                );
-            }
-            _ => unreachable!(),
-        }
-    }
-
-    #[test]
-    fn rolling_throttle_defers_all_when_a_sibling_is_unhealthy() {
-        // INVARIANT (b): mid-rollout. r0 was already recreated (now matches
-        // desired, so it is NOT a candidate) but is still STARTING (unhealthy);
-        // r1 and r2 are still drifted (running, img:OLD → rollout candidates).
-        // Because the non-candidate sibling r0 is unhealthy, ALL rollout recreates
-        // for the spec are deferred this tick → ZERO Recreates. This is the gate
-        // that makes the rollout wait for the previous replica to come back up.
-        let desired_all = vec![desired_replica(0), desired_replica(1), desired_replica(2)];
-        let actual = vec![
-            actual_replica(0, "running", "img:1"), // already updated, matches
-            actual_replica(1, "running", "img:OLD"), // drifted → candidate
-            actual_replica(2, "running", "img:OLD"), // drifted → candidate
-        ];
-        let raw = diff_desired_vs_actual(&desired_all, &actual, "rise", &no_protected());
-        // r0 (the just-recreated, non-candidate sibling) is still unhealthy.
-        let mut healthy = all_healthy(&actual);
-        healthy.insert(actual[0].identity().unwrap(), false);
-        let throttled = filter_rolling_actions(raw, &actual, &healthy);
-        assert_eq!(
-            throttled
-                .iter()
-                .filter(|a| matches!(a, ReconcileAction::Recreate { .. }))
-                .count(),
-            0,
-            "an unhealthy (still-starting) sibling defers ALL rollout recreates"
-        );
-    }
-
-    #[test]
-    fn rolling_throttle_uniform_drift_rolls_one_at_a_time() {
-        // When EVERY replica is drifted (all candidates), there are no
-        // non-candidate siblings to gate on, so the rollout begins: exactly one
-        // (r0) is recreated this tick. On the next tick r0 matches desired and
-        // becomes the sibling whose health gates the rest — yielding the
-        // one-at-a-time roll across ticks.
-        let desired_all = vec![desired_replica(0), desired_replica(1), desired_replica(2)];
-        let actual = vec![
-            actual_replica(0, "running", "img:OLD"),
-            actual_replica(1, "running", "img:OLD"),
-            actual_replica(2, "running", "img:OLD"),
-        ];
-        let raw = diff_desired_vs_actual(&desired_all, &actual, "rise", &no_protected());
-        let healthy = all_healthy(&actual);
-        let throttled = filter_rolling_actions(raw, &actual, &healthy);
-        assert_eq!(
-            throttled
-                .iter()
-                .filter(|a| matches!(a, ReconcileAction::Recreate { .. }))
-                .count(),
-            1,
-            "uniform drift still rolls one replica at a time"
-        );
-    }
-
-    #[test]
-    fn rolling_throttle_recovery_recreate_passes_even_with_unhealthy_sibling() {
-        // INVARIANT (c): a CRASHED (state != running) replica's Recreate is a
-        // recovery — it passes UNTHROTTLED even though a sibling is unhealthy.
-        // Here r0 has exited; r1 is running-drifted-unhealthy. Only r0's recovery
-        // recreate is admitted (r1's rollout recreate is gated out).
-        let desired_all = vec![desired_replica(0), desired_replica(1)];
-        let actual = vec![
-            actual_replica(0, "exited", "img:1"),    // crashed → recovery
-            actual_replica(1, "running", "img:OLD"), // running-drifted → rollout
-        ];
-        let raw = diff_desired_vs_actual(&desired_all, &actual, "rise", &no_protected());
-        // The crashed r0 is (realistically) unhealthy; it is r1's only sibling, so
-        // r1's rollout recreate is GATED OUT. r0's recovery recreate still passes.
-        let mut healthy = all_healthy(&actual);
-        healthy.insert(actual[0].identity().unwrap(), false);
-        let throttled = filter_rolling_actions(raw, &actual, &healthy);
-        let recreates: Vec<&ReconcileAction> = throttled
-            .iter()
-            .filter(|a| matches!(a, ReconcileAction::Recreate { .. }))
-            .collect();
-        assert_eq!(recreates.len(), 1, "only the recovery recreate passes");
-        match recreates[0] {
-            ReconcileAction::Recreate { name, .. } => {
-                assert!(
-                    name.contains("_r0_"),
-                    "recovery (crashed) replica passes: {name}"
-                );
-            }
-            _ => unreachable!(),
-        }
-    }
-
-    #[test]
-    fn rolling_throttle_creates_and_removes_always_pass() {
-        // INVARIANT (d): Create (scale-up) and Remove (scale-down/GC) actions are
-        // never throttled — capacity changes apply immediately. Mix a Create
-        // (r2 missing), a Remove (surplus r3), and a running-drifted r0 (rollout)
-        // with r1 healthy. The Create + Remove pass; exactly one rollout recreate
-        // passes too (r0, siblings healthy).
-        let desired_all = vec![desired_replica(0), desired_replica(1), desired_replica(2)];
-        let actual = vec![
-            actual_replica(0, "running", "img:OLD"), // rollout recreate
-            actual_replica(1, "running", "img:1"),   // matches → no action
-            actual_replica(3, "running", "img:1"),   // surplus → Remove
-        ];
-        let raw = diff_desired_vs_actual(&desired_all, &actual, "rise", &no_protected());
-        let healthy = all_healthy(&actual);
-        let throttled = filter_rolling_actions(raw, &actual, &healthy);
-        assert!(
-            throttled.iter().any(
-                |a| matches!(a, ReconcileAction::Create { name, .. } if name.contains("_r2_"))
-            ),
-            "scale-up Create passes"
-        );
-        assert!(
-            throttled
-                .iter()
-                .any(|a| matches!(a, ReconcileAction::Remove { id, .. } if id == "cid-r3")),
-            "scale-down Remove passes"
-        );
-        assert_eq!(
-            throttled
-                .iter()
-                .filter(|a| matches!(a, ReconcileAction::Recreate { .. }))
-                .count(),
-            1,
-            "exactly one rollout recreate (r0) admitted"
-        );
-    }
-
-    #[test]
     fn clamp_replicas_defaults_and_clamps() {
         assert_eq!(clamp_replicas(None), 1, "unset → 1");
         assert_eq!(clamp_replicas(Some(0)), 1, "0 → floored to 1");
@@ -3678,355 +1764,6 @@ mod tests {
             clamp_replicas(Some(MAX_REPLICAS + 100)),
             MAX_REPLICAS,
             "above max → clamped to MAX_REPLICAS"
-        );
-    }
-
-    #[test]
-    fn build_controller_metadata_aggregates_across_replicas() {
-        // 3 replicas: 2 running+ready, 1 exited → desired=3, current=2, ready=2,
-        // and three pod entries (one per replica container).
-        let pods = vec![
-            (
-                "rise_myapp_default_d_app_r0_g1".to_string(),
-                Some(inspected_running()),
-            ),
-            (
-                "rise_myapp_default_d_app_r1_g1".to_string(),
-                Some(inspected_running()),
-            ),
-            (
-                "rise_myapp_default_d_app_r2_g1".to_string(),
-                Some(inspected_exited(1)),
-            ),
-        ];
-        // Not fully ready (one replica down).
-        let meta = build_controller_metadata(&pods, &DeploymentStatus::Deploying, false);
-        let ps = &meta["pod_status"];
-        assert_eq!(ps["desired_replicas"], 3);
-        assert_eq!(ps["current_replicas"], 2);
-        assert_eq!(ps["ready_replicas"], 0); // is_ready=false → no container ready
-        assert_eq!(ps["pods"].as_array().unwrap().len(), 3);
-        assert_eq!(meta["health"]["healthy"], false);
-    }
-
-    // ── rolling in-rotation decision (pure) ────────────────────────────────
-
-    #[test]
-    fn routable_recreate_follows_is_active() {
-        use crate::server::settings::CutoverStrategy;
-        assert!(routable_for(CutoverStrategy::Recreate, true));
-        assert!(!routable_for(CutoverStrategy::Recreate, false));
-    }
-
-    #[test]
-    fn routable_health_rolling_always_true() {
-        use crate::server::settings::CutoverStrategy;
-        assert!(routable_for(CutoverStrategy::HealthRolling, true));
-        assert!(
-            routable_for(CutoverStrategy::HealthRolling, false),
-            "rolling mode is routable even when not active"
-        );
-    }
-
-    #[test]
-    fn rotation_health_check_up_is_in_rotation() {
-        // HC configured, Traefik API available, server UP → in rotation.
-        assert_eq!(
-            rolling_rotation_decision(true, true, true, Some(true)),
-            RotationDecision::InRotation
-        );
-    }
-
-    #[test]
-    fn rotation_health_check_down_is_not_in_rotation() {
-        // HC configured, API available, server DOWN → NOT in rotation, with the
-        // Traefik-specific reason (distinct from the probe reasons).
-        match rolling_rotation_decision(true, true, true, Some(false)) {
-            RotationDecision::NotInRotation(reason) => {
-                assert!(
-                    reason.contains("DOWN"),
-                    "reason should mention Traefik server DOWN, got: {reason}"
-                );
-            }
-            other => panic!("expected NotInRotation, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn rotation_health_check_absent_server_is_not_in_rotation() {
-        // HC configured, API available, but the server URL isn't in the
-        // serverStatus map yet → not in rotation, distinct reason.
-        match rolling_rotation_decision(true, true, true, None) {
-            RotationDecision::NotInRotation(reason) => {
-                assert!(reason.contains("serverStatus"), "got: {reason}");
-            }
-            other => panic!("expected NotInRotation, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn rotation_no_health_check_running_is_in_rotation() {
-        // No HC → ready-when-running: a running container is in rotation
-        // regardless of any Traefik signal.
-        assert_eq!(
-            rolling_rotation_decision(false, true, false, None),
-            RotationDecision::InRotation
-        );
-        assert_eq!(
-            rolling_rotation_decision(false, true, true, Some(false)),
-            RotationDecision::InRotation,
-            "no-HC + running ignores Traefik serverStatus"
-        );
-    }
-
-    #[test]
-    fn rotation_no_health_check_not_running_is_not_in_rotation() {
-        match rolling_rotation_decision(false, false, false, None) {
-            RotationDecision::NotInRotation(reason) => {
-                assert!(reason.contains("not running"), "got: {reason}");
-            }
-            other => panic!("expected NotInRotation, got {other:?}"),
-        }
-    }
-
-    // ── service-name derivation (matches the labels) ───────────────────────
-
-    #[test]
-    fn service_names_for_spec_single_route_is_bare_base() {
-        use crate::server::deployment::models::{ContainerSpec, RouteSpec};
-        let spec = ContainerSpec {
-            name: "app".to_string(),
-            image: None,
-            port: Some(8080),
-            replicas: None,
-            cpu: None,
-            memory: None,
-            env_overrides: vec![],
-            health_check: None,
-        };
-        let routes = vec![RouteSpec {
-            path: "/".to_string(),
-            container: "app".to_string(),
-        }];
-        // A single-route container uses the bare base name — exactly what the
-        // labels stamp (`group_service_names` over the matching DesiredContainer).
-        assert_eq!(
-            service_names_for_spec("myapp", "default", &spec, &routes),
-            vec!["myapp-default-app".to_string()]
-        );
-    }
-
-    #[test]
-    fn service_names_for_spec_multi_route_uses_per_route_indices() {
-        use crate::server::deployment::models::{ContainerSpec, RouteSpec};
-        let spec = ContainerSpec {
-            name: "api".to_string(),
-            image: None,
-            port: Some(8080),
-            replicas: None,
-            cpu: None,
-            memory: None,
-            env_overrides: vec![],
-            health_check: None,
-        };
-        // Two routes for `api` (+ one for another container, which is ignored).
-        let routes = vec![
-            RouteSpec {
-                path: "/".to_string(),
-                container: "api".to_string(),
-            },
-            RouteSpec {
-                path: "/api/v1".to_string(),
-                container: "api".to_string(),
-            },
-            RouteSpec {
-                path: "/".to_string(),
-                container: "other".to_string(),
-            },
-        ];
-        // Per-route services, longest path-prefix first (-0 = /api/v1, -1 = /),
-        // matching the renderer's `{base}-{idx}` services for a multi-route
-        // container. The bare base is NOT queried (it 404s in the Traefik API).
-        assert_eq!(
-            service_names_for_spec("myapp", "default", &spec, &routes),
-            vec![
-                "myapp-default-api-0".to_string(),
-                "myapp-default-api-1".to_string()
-            ]
-        );
-    }
-
-    #[test]
-    fn service_names_for_spec_derivation_matches_container_builder() {
-        // The reconciler-side derivation must agree with the builder-side
-        // `group_service_names` (the labels) for the same routes — single AND
-        // multi-route — so the lookup never drifts from what's stamped.
-        use crate::server::deployment::models::{ContainerSpec, RouteSpec};
-        use container_builder::{group_service_names, DesiredRoute};
-
-        let mk_desired = |container: &str, paths: &[&str]| {
-            let mut d = desired(container, "img:1", "h1");
-            d.routes = paths
-                .iter()
-                .map(|p| DesiredRoute {
-                    hosts: vec!["myapp.rise.dev".to_string()],
-                    path_prefix: Some(p.to_string()),
-                })
-                .collect();
-            d
-        };
-        let mk_spec_routes = |container: &str, paths: &[&str]| {
-            let spec = ContainerSpec {
-                name: container.to_string(),
-                image: None,
-                port: Some(8080),
-                replicas: None,
-                cpu: None,
-                memory: None,
-                env_overrides: vec![],
-                health_check: None,
-            };
-            let routes: Vec<RouteSpec> = paths
-                .iter()
-                .map(|p| RouteSpec {
-                    path: p.to_string(),
-                    container: container.to_string(),
-                })
-                .collect();
-            (spec, routes)
-        };
-
-        for paths in [&["/"][..], &["/", "/api/v1"][..], &["/", "/a", "/bb"][..]] {
-            let (spec, routes) = mk_spec_routes("app", paths);
-            let from_spec = service_names_for_spec("myapp", "default", &spec, &routes);
-            let from_labels = group_service_names(&mk_desired("app", paths));
-            assert_eq!(
-                from_spec, from_labels,
-                "service-name derivation must match the labels for paths {paths:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn service_names_for_spec_worker_has_no_services() {
-        use crate::server::deployment::models::ContainerSpec;
-        let worker = ContainerSpec {
-            name: "worker".to_string(),
-            image: None,
-            port: None,
-            replicas: None,
-            cpu: None,
-            memory: None,
-            env_overrides: vec![],
-            health_check: None,
-        };
-        assert!(service_names_for_spec("myapp", "default", &worker, &[]).is_empty());
-    }
-
-    // ── replica_ready (pure readiness selection) ───────────────────────────
-
-    #[test]
-    fn replica_ready_rolling_server_up_is_ready() {
-        // Rolling + HC + Traefik UP → Ready, no probe needed.
-        assert_eq!(
-            replica_ready(true, true, true, true, Some(true), None),
-            ReadyVerdict::Ready
-        );
-    }
-
-    #[test]
-    fn replica_ready_rolling_server_down_is_not_ready() {
-        match replica_ready(true, true, true, true, Some(false), None) {
-            ReadyVerdict::NotReady(reason) => assert!(reason.contains("DOWN"), "got: {reason}"),
-            other => panic!("expected NotReady, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn replica_ready_rolling_absent_server_is_not_ready() {
-        match replica_ready(true, true, true, true, None, None) {
-            ReadyVerdict::NotReady(reason) => {
-                assert!(reason.contains("serverStatus"), "got: {reason}")
-            }
-            other => panic!("expected NotReady, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn replica_ready_rolling_fallback_needs_probe_then_folds_result() {
-        // Rolling + HC but Traefik API unavailable → fall back to Rise's probe.
-        // With no probe result yet → NeedsProbe; folding the result in resolves.
-        assert_eq!(
-            replica_ready(true, true, true, false, None, None),
-            ReadyVerdict::NeedsProbe
-        );
-        assert_eq!(
-            replica_ready(true, true, true, false, None, Some(true)),
-            ReadyVerdict::Ready
-        );
-        assert!(matches!(
-            replica_ready(true, true, true, false, None, Some(false)),
-            ReadyVerdict::NotReady(_)
-        ));
-    }
-
-    #[test]
-    fn replica_ready_rolling_no_health_check_is_ready_when_running() {
-        // Rolling, no HC → ready-when-running; Traefik signal ignored.
-        assert_eq!(
-            replica_ready(true, false, true, false, None, None),
-            ReadyVerdict::Ready
-        );
-        assert!(matches!(
-            replica_ready(true, false, false, false, None, None),
-            ReadyVerdict::NotReady(_)
-        ));
-    }
-
-    #[test]
-    fn replica_ready_non_rolling_uses_probe_when_health_path() {
-        // Non-rolling + HC: readiness IS the probe result.
-        assert_eq!(
-            replica_ready(false, true, true, false, None, None),
-            ReadyVerdict::NeedsProbe
-        );
-        assert_eq!(
-            replica_ready(false, true, true, false, None, Some(true)),
-            ReadyVerdict::Ready
-        );
-        assert!(matches!(
-            replica_ready(false, true, true, false, None, Some(false)),
-            ReadyVerdict::NotReady(_)
-        ));
-    }
-
-    #[test]
-    fn replica_ready_non_rolling_no_health_path_gates_on_running() {
-        // Non-rolling, no HC → ready-when-running with the disabled-probe reason.
-        assert_eq!(
-            replica_ready(false, false, true, false, None, None),
-            ReadyVerdict::Ready
-        );
-        match replica_ready(false, false, false, false, None, None) {
-            ReadyVerdict::NotReady(reason) => {
-                assert!(reason.contains("not running"), "got: {reason}")
-            }
-            other => panic!("expected NotReady, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn rotation_health_check_api_unavailable_falls_back_to_probe() {
-        // HC configured but Traefik API unset/errored → MIRROR with Rise's own
-        // probe (graceful degradation). The `running`/`server_up` inputs are
-        // ignored in this branch.
-        assert_eq!(
-            rolling_rotation_decision(true, true, false, None),
-            RotationDecision::FallBackToProbe
-        );
-        assert_eq!(
-            rolling_rotation_decision(true, false, false, Some(true)),
-            RotationDecision::FallBackToProbe
         );
     }
 }
